lazy import in aws client fixtures
