+++ tests/conftest.py
@@ -52,3 +52,49 @@ def pytest_collection_modifyitems(config, items):
     for item in items:
         if "skip_offline" in item.keywords:
             item.add_marker(skip_offline)
+
+
+@pytest.fixture(scope="session")
+def aws_session():
+    """
+    This fixture returns the Boto Session instance for testing.
+    """
+    from localstack.testing.aws.util import base_aws_session
+
+    return base_aws_session()
+
+
+@pytest.fixture(scope="session")
+def aws_client_factory(aws_session):
+    """
+    This fixture returns a client factory for testing.
+
+    Use this fixture if you need to use custom endpoint or Boto config.
+    """
+    from localstack.testing.aws.util import base_aws_client_factory
+
+    return base_aws_client_factory(aws_session)
+
+
+@pytest.fixture(scope="session")
+def aws_client(aws_client_factory):
+    """
+    This fixture can be used to obtain Boto clients for testing.
+
+    The clients are configured with the primary testing credentials.
+    """
+    from localstack.testing.aws.util import primary_testing_aws_client
+
+    return primary_testing_aws_client(aws_client_factory)
+
+
+@pytest.fixture(scope="session")
+def secondary_aws_client(aws_client_factory):
+    """
+    This fixture can be used to obtain Boto clients for testing.
+
+    The clients are configured with the secondary testing credentials.
+    """
+    from localstack.testing.aws.util import secondary_testing_aws_client
+
+    return secondary_testing_aws_client(aws_client_factory)
