--- tests/fixtures.py
@@ -1,46 +1,0 @@
-import pytest
-
-from localstack.testing.aws.util import (
-    base_aws_client_factory,
-    base_aws_session,
-    primary_testing_aws_client,
-    secondary_testing_aws_client,
-)
-
-
-@pytest.fixture(scope="session")
-def aws_session():
-    """
-    This fixture returns the Boto Session instance for testing.
-    """
-    return base_aws_session()
-
-
-@pytest.fixture(scope="session")
-def aws_client_factory(aws_session):
-    """
-    This fixture returns a client factory for testing.
-
-    Use this fixture if you need to use custom endpoint or Boto config.
-    """
-    return base_aws_client_factory(aws_session)
-
-
-@pytest.fixture(scope="session")
-def aws_client(aws_client_factory):
-    """
-    This fixture can be used to obtain Boto clients for testing.
-
-    The clients are configured with the primary testing credentials.
-    """
-    return primary_testing_aws_client(aws_client_factory)
-
-
-@pytest.fixture(scope="session")
-def secondary_aws_client(aws_client_factory):
-    """
-    This fixture can be used to obtain Boto clients for testing.
-
-    The clients are configured with the secondary testing credentials.
-    """
-    return secondary_testing_aws_client(aws_client_factory)
